Added additional data to first transformation feature.
Additional data makes it so I can't just return expected value from
Nate template engine i.e. it has to do some real work of some sort.

Put in a hack in template to deal with Nokogiri adding in newlines.
Will have to be dealt with at a later point in time but all is green
right now.

Specs need be added.