Made encoders more robust.
Previously, as part of the general proof of concept, we just required in the encoder method
which was prone to lots of bugs in real world usage. Moved to a slightly more intelligent
strategy pattern in case someone actually puts this into real world usage even though
at this stage it is all basic proof of concept prototype stuff w/ a ton of nokogiri
related issues to be explored.